requirements.txt
alembic==1.15.2
Flask==2.3.2
Flask-Cors==4.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.38
redis==5.0.1
flask-jwt-extended==4.5.3
marshmallow==3.20.1
celery==5.3.4
redis==5.0.1
